requirements.txt
# Libraries-modules pre_written
boto3==1.34.158
botocore==1.34.158
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
python-jenkins